Move Redux store and reducer into wrapping provider component
