+++ src/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import Store from './Store';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Store />
   </React.StrictMode>,
   document.getElementById('root')
 );
